requirements.txt
pandas
scikit-learn
flask
pytesseract
Pillow
spacy
pdf2image
openpyxl
